+++ src/main/java/com/ontotext/refine/cli/export/rdf/AllowedRdfResultFormat.java
@@ -1,9 +1,7 @@
 package com.ontotext.refine.cli.export.rdf;
 
-import com.ontotext.refine.client.command.rdf.ResultFormat;
 import java.util.Arrays;
 import java.util.Iterator;
-import java.util.Set;
 
 /**
  * Provides the values for the allowed RDF result formats and the completion candidates for the RDF
@@ -13,12 +11,10 @@ import java.util.Set;
  */
 public class AllowedRdfResultFormat implements Iterable<String> {
 
-  private static final Set<ResultFormat> UNSUPPORTED = Set.of(ResultFormat.HDT, ResultFormat.RDFA);
-
   @Override
   public Iterator<String> iterator() {
-    return Arrays.stream(ResultFormat.values())
-        .filter(value -> !UNSUPPORTED.contains(value))
+    return Arrays
+        .stream(RdfResultFormats.values())
         .map(value -> value.toString().toLowerCase())
         .iterator();
   }